Set CommitOnSuccess to true for some commands
These commands are expected to return 0 to mean that a commit should be added.
Reportmail and Help are two commands whose successful execution should not signal
the creation of a new commit.

Some commands have had NOCOMMIT=true removed, since the new CommitOnSuccess is set
to False accordingly.

Other commands that are now understood to be no-commit have had their "exit 42" removed.