Remove RegionServer prereq for HBase 1.1.4
With the new version of HBase the prereq was introducing a dependency
loop in which the master rest service would not respond until the region
servers were online, but the region servers could not start until the
master rest service was responding.  Deleting the prereq allows
everything to start normally.