Add zenmapper service definition
This must be added to resmgr & resmgr.lite now that the Layer2 ZenPack
is installed by default. These service definitions and configuration
files are suited to version 1.3.5 of the Layer2 ZenPack because that is
what is currently slated to ship with Thebe.

The next major version release of the Layer2 ZenPack will move the
zenmapper service under the Zenoss/ path, but for now it must be at the
root level to match what ships in the ZenPack.

Fixes ZPS-2350.